requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.0
fsspec==2025.5.1
gotrue==2.12.0
gunicorn==23.0.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
openai==1.84.0
voyageai==0.2.3
packaging==24.2
pillow==11.2.1
postgrest==1.0.2
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-settings==2.9.1
PyJWT==2.10.1
PyMuPDF==1.26.1
PyPDF2==3.0.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
realtime==2.4.3
redis==6.2.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
storage3==0.11.3
supabase==2.15.2
supafunc==0.9.4
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tqdm==4.67.1
Unidecode==1.4.0
urllib3==2.4.0
Werkzeug==3.1.3
huggingface_hub>=0.16.4
sentence-transformers==2.7.0
torch==2.0.1+cpu --find-links https://download.pytorch.org/whl/torch_stable.html
transformers==4.35.0